chat_9haziran/requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
numpy>=1.24.0
torch>=2.0.0
transformers>=4.21.0